requirements.txt
tensorflow-gpu==2.4.0
numpy==1.19.2
pandas==1.1.3
scikit-learn==0.23.2
tqdm==4.50.2
click==7.1.2
matplotlib==3.3.2